fix(serial_driver): Add connection helpers
Add a connect wait to allow for wait only after connect.
This is needed since Arduinos autoreset on connect and sending too early will lock the bootloader.
The timeout was also cleaned up since ERR can be valid.
Also get rid of the auto flush.
Auto flush causes problems in some boards and should be done in a higher layer.